+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Dependencies: bump textual from 0.85.0 to 0.85.1 #35
+- Hide statuses with zero torrents from state panel [#36](https://github.com/anlar/tewi/issues/36)
+- Dependencies: bump textual from 0.85.0 to 0.85.1 [#35](https://github.com/anlar/tewi/issues/35)
 
 ### Removed